+++ releasenotes.md
@@ -5,7 +5,8 @@
 - refactor: handling defaults
 - #19 added --batchsize CLI parameter
 - #23 added --listconfig CLI parameter
-- added validation for dlc, executable, workdir, srcroot & batchsize
+- #20 added validation for dlc, executable, workdir, srcroot & batchsize
+- #24 added --workdir and --srcroot CLI parameters
   
 `1.2.0`
 - #14 added --targetdir (-t) CLI parameter